requirements.txt
gunicorn
certifi==2020.6.20
cffi
click==7.1.2
cryptography
dnspython
dominate==2.4.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-WTF==0.14.3
idna
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe
pycparser
six
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
numpy==1.19.3
Pillow==7.2.0
requests==2.24.0